+++ CHANGELOG.md
@@ -32,4 +32,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Basic README file.
 - Automatic publishing of tagged versions.
 - Caching of experimental parameters and chemical descriptors.
+- Choices for acquisition functions and their usage with arbitrary surrogate models.